Refactored code file structure, minor fixes
